+++ packages/widget/src/ui/Dialog/DialogContent.tsx
@@ -11,28 +11,29 @@ interface Props extends PropsWithChildren {
 }
 
 export function DialogContent({ children, onInteractOutside }: Props) {
-  const { open, container } = useContext(DialogContext);
+  const { open } = useContext(DialogContext);
 
   if (!open) return null;
 
   return (
-    <Dialog.Portal>
-      <Dialog.Overlay className="bg-blackA6 data-[state=open]:animate-overlayShow fixed inset-0 font-jost" />
-      <DialogWrapper
+    <DialogWrapper>
+      <Dialog.Overlay className="bg-blackA6 data-[state=open]:animate-overlayShow fixed inset-0 font-diatype z-10" />
+      <Dialog.Content
         className={cn(
           'data-[state=open]:animate-contentShow fixed top-[50%] left-[50%]',
           'w-[90vw] max-w-[450px] max-h-[820px] h-[90vh] rounded-xl',
-          'translate-x-[-50%] translate-y-[-50%] bg-white shadow-[hsl(206_22%_7%_/_35%)_0px_10px_38px_-10px,_hsl(206_22%_7%_/_20%)_0px_10px_20px_-15px]'
+          'translate-x-[-50%] translate-y-[-50%] bg-white shadow-[hsl(206_22%_7%_/_35%)_0px_10px_38px_-10px,_hsl(206_22%_7%_/_20%)_0px_10px_20px_-15px]',
+          'z-20'
         )}
         onInteractOutside={onInteractOutside}
       >
         {children}
-      </DialogWrapper>
-    </Dialog.Portal>
+      </Dialog.Content>
+    </DialogWrapper>
   );
 }
 
-const DialogWrapper = styled(Dialog.Content)`
+const DialogWrapper = styled.div`
   background-color: ${(props) => props.theme.primary.backgroundColor};
   color: ${(props) => props.theme.primary.textColor};
 `;