+++ packages/widget/src/ui/SettingsDialog/index.tsx
@@ -3,7 +3,6 @@ import * as Dialog from '@radix-ui/react-dialog';
 
 import { AdaptiveLink } from '../AdaptiveLink';
 import { GasSetting } from './GasSetting';
-import { PurgeSetting } from './PurgeSetting';
 import { SaveIndicator } from './SaveIndicator';
 import { SlippageSetting } from './SlippageSetting';
 import { useDisclosureKey } from '../../store/disclosures';
@@ -16,7 +15,7 @@ export const SettingsDialog = () => {
       <Dialog.Overlay className="bg-blackA6 data-[state=open]:animate-overlayShow fixed inset-0" />
       <Dialog.Content
         className={cn(
-          'font-jost data-[state=open]:animate-contentShow fixed top-[50%] left-[50%]',
+          'font-diatype data-[state=open]:animate-contentShow fixed top-[50%] left-[50%]',
           'w-[90vw] max-w-[450px] max-h-[820px] h-[90vh] rounded-xl',
           'translate-x-[-50%] translate-y-[-50%] bg-white shadow-[hsl(206_22%_7%_/_35%)_0px_10px_38px_-10px,_hsl(206_22%_7%_/_20%)_0px_10px_20px_-15px]'
         )}