+++ packages/widget/src/ui/SimpleTooltip.tsx
@@ -2,7 +2,6 @@ import * as Tooltip from '@radix-ui/react-tooltip';
 import { ComponentPropsWithoutRef, ReactNode } from 'react';
 import { cn } from '../utils/ui';
 import { useSwapWidgetUIStore } from '../store/swap-widget';
-import { css } from '@emotion/css';
 
 type Props = Tooltip.TooltipProps & {
   type?: 'default' | 'warning' | 'brand';
@@ -27,43 +26,41 @@ export const SimpleTooltip = (props: Props) => {
   return (
     <Tooltip.Root {...tooltipProps}>
       <Tooltip.Trigger asChild>{children}</Tooltip.Trigger>
-      <Tooltip.Portal>
-        <Tooltip.Content
-          sideOffset={4}
-          {..._content}
+      <Tooltip.Content
+        sideOffset={4}
+        {..._content}
+        className={cn(
+          'rounded-md bg-white px-4 py-2 leading-none',
+          'select-none shadow shadow-neutral-500/50',
+          'text-sm font-diatype',
+          'animate-slide-up-and-fade',
+          type === 'warning' && `bg-[#fbeef1]`,
+          type === 'warning' && 'font-medium',
+          type === 'brand' && `text-white`,
+          _content?.className
+        )}
+        style={{
+          backgroundColor:
+            type === 'brand'
+              ? useSwapWidgetUIStore.getState().colors.primary
+              : undefined,
+          color:
+            type === 'warning'
+              ? useSwapWidgetUIStore.getState().colors.primary
+              : undefined,
+        }}
+      >
+        {label}
+        <Tooltip.Arrow
           className={cn(
-            'rounded-md bg-white px-4 py-2 leading-none',
-            'select-none shadow shadow-neutral-500/50',
-            'text-sm font-jost',
-            'animate-slide-up-and-fade',
-            type === 'warning' && `bg-[#fbeef1]`,
-            type === 'warning' &&
-              css`
-                color: ${useSwapWidgetUIStore.getState().colors.primary};
-              `,
-            type === 'warning' && 'font-medium',
-            type === 'brand' && `text-white`,
-            type === 'brand' &&
-              css`
-                background-color: ${useSwapWidgetUIStore.getState().colors
-                  .primary};
-              `,
-            _content?.className
+            'fill-white drop-shadow',
+            type === 'warning' && 'fill-[#fbeef1]'
           )}
-        >
-          {label}
-          <Tooltip.Arrow
-            className={cn(
-              'fill-white drop-shadow',
-              type === 'warning' && 'fill-[#fbeef1]',
-              type === 'brand' &&
-                css`
-                  fill: ${useSwapWidgetUIStore.getState().colors.primary};
-                `
-            )}
-          />
-        </Tooltip.Content>
-      </Tooltip.Portal>
+          style={{
+            fill: useSwapWidgetUIStore.getState().colors.primary,
+          }}
+        />
+      </Tooltip.Content>
     </Tooltip.Root>
   );
 };