+++ packages/widget/src/ui/Widget.tsx
@@ -25,22 +25,52 @@ import { SpinnerIcon } from './Icon/SpinnerIcon';
 import { useSwapWidgetUIStore } from '../store/swap-widget';
 import { css } from '@emotion/css';
 import { CraftedBySkip } from './CraftedBySkip';
+import { useFixRadixUiWheelEvent } from '../hooks/use-fix-radix-ui-wheel-event';
 import { ThemeProvider, styled } from 'styled-components';
 import { Theme, defaultTheme } from './theme';
 
-type SwapWidgetProps = React.HTMLAttributes<HTMLDivElement> & {
+type SwapWidgetUIProps = Pick<
+  React.HTMLAttributes<HTMLDivElement>,
+  'className' | 'style'
+> & {
   theme?: Theme;
 };
 
 export const SwapWidgetUI = ({
   className,
+  style,
   theme,
-  ...divProps
-}: SwapWidgetProps) => {
+}: SwapWidgetUIProps) => {
+  useFixRadixUiWheelEvent();
+
   useEffect(() => void disclosure.rehydrate(), []);
 
   const { openWalletModal } = useWalletModal();
-  const { data: chains } = useChains();
+
+  const filter = useSwapWidgetUIStore((state) => state.filter);
+  const sourceChainIDs = filter?.source
+    ? Object.keys(filter.source)
+    : undefined;
+  const destinationChainIDs = filter?.destination
+    ? Object.keys(filter.destination)
+    : undefined;
+
+  const { data: sourceChains } = useChains({
+    select: (chains) => {
+      if (!sourceChainIDs) return chains;
+      return chains?.filter((chain) => {
+        return sourceChainIDs?.includes(chain.chainID);
+      });
+    },
+  });
+  const { data: destinationChains } = useChains({
+    select: (chains) => {
+      if (!destinationChainIDs) return chains;
+      return chains?.filter((chain) => {
+        return destinationChainIDs?.includes(chain.chainID);
+      });
+    },
+  });
 
   const {
     amountIn,
@@ -101,10 +131,10 @@ export const SwapWidgetUI = ({
         <Tooltip.Provider delayDuration={0} disableHoverableContent>
           <WidgetContainer
             className={cn(
-              'space-y-4 font-jost relative p-4 bg-white',
+              'space-y-4 font-diatype relative p-4 bg-white',
               className
             )}
-            {...divProps}
+            style={style}
           >
             <div className="flex h-8 items-center">
               <p className="text-2xl font-semibold">From</p>
@@ -142,7 +172,7 @@ export const SwapWidgetUI = ({
                 amountUSD={route?.usdAmountIn}
                 asset={sourceAsset}
                 chain={sourceChain}
-                chains={chains ?? []}
+                chains={sourceChains ?? []}
                 onAmountChange={onSourceAmountChange}
                 onAmountMax={onSourceAmountMax}
                 onAssetChange={onSourceAssetChange}
@@ -182,7 +212,7 @@ export const SwapWidgetUI = ({
                 diffPercentage={usdDiffPercent}
                 asset={destinationAsset}
                 chain={destinationChain}
-                chains={chains ?? []}
+                chains={destinationChains ?? []}
                 onAmountChange={onDestinationAmountChange}
                 onAssetChange={onDestinationAssetChange}
                 onChainChange={onDestinationChainChange}
@@ -217,20 +247,18 @@ export const SwapWidgetUI = ({
             {route && !routeLoading && numberOfTransactions > 1 && (
               <div className="flex w-full items-center justify-center space-x-2 text-sm font-medium uppercase">
                 <div
-                  className={cn(
-                    'relative rounded-full p-[4px]',
-                    css`
-                      background-color: ${useSwapWidgetUIStore.getState().colors
-                        .primary};
-                    `
-                  )}
+                  className={cn('relative rounded-full p-[4px]')}
+                  style={{
+                    backgroundColor:
+                      useSwapWidgetUIStore.getState().colors.primary,
+                  }}
                 >
                   <div
                     className={cn(
                       'absolute h-6 w-6 animate-ping rounded-full',
                       css`
                         background-color: ${useSwapWidgetUIStore.getState()
-                          .colors.primary};
+                          .colors.primary} !important;
                       `
                     )}
                   />
@@ -261,12 +289,12 @@ export const SwapWidgetUI = ({
                 className={cn(
                   'w-full rounded-md py-4 font-semibold text-white outline-none transition-[opacity,transform]',
                   'disabled:cursor-not-allowed disabled:opacity-75',
-                  'enabled:hover:rotate-1 enabled:hover:scale-105',
-                  css`
-                    background-color: ${useSwapWidgetUIStore.getState().colors
-                      .primary};
-                  `
+                  'enabled:hover:rotate-1 enabled:hover:scale-105'
                 )}
+                style={{
+                  backgroundColor:
+                    useSwapWidgetUIStore.getState().colors.primary,
+                }}
                 disabled={!sourceChain}
                 onClick={async () => {
                   if (sourceChain && !srcAccount?.isWalletConnected) {