+++ packages/widget/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @skip-go/widget
 
+## 1.2.3
+
+### Patch Changes
+
+- 1e173c3: Remove extra SwapWidgetProvider wrapping the web-component
+
+## 1.2.2
+
+### Patch Changes
+
+- 7240d7b: Update widget README
+
+## 1.2.1
+
+### Patch Changes
+
+- 4939650: source/destination chains/assets filtering
+
+## 1.2.0
+
+### Minor Changes
+
+- 6c4e611: Move widget to within shadow DOM
+
+## 1.1.1
+
+### Patch Changes
+
+- 6c044e1: Use latest core version and update readme
+
+## 1.1.0
+
+### Minor Changes
+
+- 2c7e965: change url
+
+## 1.0.0
+
+### Major Changes
+
+- 0380db6: change fonts and logo
+
+## 0.1.1
+
+### Patch Changes
+
+- 1e84585: prefixwrap css styles
+
+## 0.1.0
+
+### Minor Changes
+
+- e05771b: Add web-component
+
 ## 0.0.1
 
 ### Patch Changes